+++ packages/components/src/custom-select-control/index.tsx
@@ -7,10 +7,10 @@ import clsx from 'clsx';
 /**
  * Internal dependencies
  */
-import _CustomSelect from '../custom-select';
-import CustomSelectItem from '../item';
-import type { LegacyCustomSelectProps } from '../types';
-import * as Styled from '../styles';
+import _CustomSelect from '../custom-select-control-v2/custom-select';
+import CustomSelectItem from '../custom-select-control-v2/item';
+import type { LegacyCustomSelectProps } from '../custom-select-control-v2/types';
+import * as Styled from '../custom-select-control-v2/styles';
 
 function useDeprecatedProps( {
 	__experimentalShowSelectedHint,
